python script
I have tried 20 lines operation in table title and it worked on my computer.

But I need to have a discussion on schema because I quite not understand some of tables.

Moreover, my code is easy enough to read, you can help me with rest table if you got time